finished permissions implementation

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -4,6 +4,7 @@ const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const mutations = {
   async createItem(parent, args, context, info) {
@@ -49,7 +50,19 @@ const mutations = {
 
   async deleteItem(parent, args, context, info) {
     const where = { id: args.id };
-    const item = await context.db.query.item({ where }, `{ id, title}`);
+    const item = await context.db.query.item(
+      { where },
+      `{ id, title, user { id }}`
+    );
+    const ownsItem = item.user.id === context.request.userId;
+    const hasPermissions = context.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You do not have permissions!");
+    }
+
     return context.db.mutation.deleteItem({ where }, info);
   },
 
@@ -181,6 +194,38 @@ const mutations = {
     });
     // 8. Return the new user
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in!");
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
